Added a new message type "DIRECTORY" to the message type class
Added this so that the client can request the user directory from the server by passing in a message of message type "DIRECTORY" and the server will send back a message object which will contain all the user names. Added the functions to the server and client to perform this, still working on it. Not completed at the moment.